services/ll_manager/requirements.txt
Flask-Migrate==4.0.7
flask-openapi3==3.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
